+++ src/Configuration.tsx
@@ -14,7 +14,7 @@ function Configuration(props: ConfigurationProps) {
 
   return (
     <div>
-      <Typography class="custom-h2" variant="h2">
+      <Typography class="custom-h2" variant="h2" color="text.primary">
         Configuration
       </Typography>
       <TableContainer>