+++ src/index.tsx
@@ -1,8 +1,8 @@
-/* @refresh reload */
+// @refresh reload
 import { render } from "solid-js/web";
-import App from "./App";
+import ThemedApp from "./ThemedApp";
 import reportWebVitals from "./reportWebVitals";
 
-render(() => <App />, document.getElementById("root")!);
+render(() => <ThemedApp />, document.getElementById("root")!);
 
 reportWebVitals(console.log);